# doc/api.yaml
openapi: 3.1.0
servers:
  - url: "/v1"
info:
  title: WASAText
  description: WASAText; "Programmazione per il Web 2025/26"'s project
  version: 0.1.0
...
security:
  - BearerAuth: []
paths:
...
  /conversations/{convId}/participants:
    post:
      summary: add a member to the group
      description: add to the array list of the participants of the group an user
      operationId: addToGroup
      tags: ["groups"]
      parameters:
        - name: convId
          description: conversation's id
          in: path
          required: true
          schema: 
            $ref: '#/components/schemas/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: array of the users'id
              properties:
                userIds:
                  description: list of the users' id present in the group chat
                  type: array
                  items:
                    $ref: '#/components/schemas/UserId'
                  minItems: 1
                  maxItems: 20
              required:
                - userIds
      responses:
        '200':
          description: Group updated successfully, with new participants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                convId: "conv_78910"
                type: "group"
                groupName: "Study Group 2026"
                participants: ["usr_123", "usr_456", "usr_789", "usr_101"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /conversations/{convId}/participants/me:              
    delete:
      description: Let the user leave a group chat
      operationId: leaveGroup
      summary: leave the group
      tags: ["groups"]
      parameters:
        - name: convId
          in: path
          required: true
          description: conversation's id
          schema: 
            $ref: '#/components/schemas/ConversationId'
      responses:
        '200':
          description: Successfully left the group. Returns the updated group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                convId: "conv_78910"
                type: "group"
                groupName: "Study Group 2026"
                participants: ["usr_123", "usr_456"]  # Assuming the user leaving was usr_789
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /conversations/{convId}/group_name:
    put:
      operationId: setGroupName
      description: the user sets the group name
      summary: set the group name
      tags: ["set", "groups"]
      parameters:
        - name: convId
          in: path
          required: true
          description: conversation's id
          schema: 
            $ref: '#/components/schemas/ConversationId'
      requestBody:
        required: true 
        content:
          application/json:
            schema:
              type: object
              description: group name to set
              properties:
                groupName:
                  description: group's name that can be modified by the users
                  type: string
                  pattern: '^[A-Za-z0-9_-]+$'
                  minLength: 1
                  maxLength: 16
              required:
                - groupName
      responses: 
        '200':
          description: group name set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                convId: "conv_78910"
                type: "group"
                groupName: "New Study Group Name"
                participants: ["usr_123", "usr_456", "usr_789"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /conversations/{convId}:
    get:
      summary: get a specific conversation
      description: returns details of a single conversation by its identifier
      operationId: getConversation
      tags: ["groups","get"]
      parameters:
        - name: convId
          description: conversation's id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ConversationId'
      responses:
        '200':
          description: conversation details
          content:
            application/json:
              example:
                convId: "conv_123456"
                type: "private"
                participants: ["usr_123", "usr_456"]
              schema:
                $ref: '#/components/schemas/Conversation'
                
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /conversations/{convId}/messages/{messageId}/forwarded:
    post:
      operationId: forwardMessage
      tags: ["forwardedMessages"]
      summary: Forward an existing message to this conversation
      description: |
        Creates a forwarded copy of an existing message within the specified conversation.
        The forwarded message retains the original content but is marked as "forwarded".
      parameters:
      - name: convId
        in: path
        required: true
        description: conversation's id
        schema:
          $ref: '#/components/schemas/ConversationId'
      - name: messageId
        in: path
        required: true
        description: message's id of the message to forward
        schema: { $ref: '#/components/schemas/MessageId' }

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload to forward a message to another conversation
              properties:
                destinationConversationId:
                  $ref: '#/components/schemas/ConversationId'
              required:
                - destinationConversationId
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                messageId: "msg_forwarded_123"
                text: "Hello! (forwarded)"
                sender: "usr_123"
                time: "2024-01-01T12:00:00Z"
                status: "sent"
                kind: "forwarded"
                reactions: []
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /conversations/{convId}/messages/{messageId}/reaction:
    parameters:
        - name: convId
          in: path
          required: true
          description: conversation's id
          schema:
            $ref: '#/components/schemas/ConversationId'
        - name: messageId
          in: path
          required: true
          description: message's id
          schema:
            $ref: '#/components/schemas/MessageId'
    put:
      tags: ["messages"]
      summary: React to a message
      description: |
        Adds or updates the user's reaction to a message.
        If the user already reacted, the old reaction is replaced with the new emoji.
        This enforces a "one reaction per user" policy (like WhatsApp/Instagram).
      operationId: commentMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Payload to set a reaction on a message
              type: object
              properties:
                emoji:
                  type: string
                  pattern: '^.+$'
                  description: The emoji to react with
                  minLength: 1
                  maxLength: 4
                  example: "👍"
              required:
                - emoji
      responses:
        '200':
          description: Reaction set successfully. Returns the updated message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                messageId: "msg_123"
                text: "Hello!"
                sender: "usr_123"
                time: "2024-01-01T12:00:00Z"
                status: "sent"
                reactions: [
                  { userId: "usr_456", emoji: "👍" }
                ]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundResponse'

    delete:
      tags: ["messages"]
      summary: Remove reaction
      description: Removes the user's reaction from the message (Un-react).
      operationId: uncommentMessage
      responses:
        '200':
          description: Reaction removed successfully. Returns the updated message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                messageId: "msg_123"
                text: "Hello!"
                sender: "usr_123"
                time: "2024-01-01T12:00:00Z"
                status: "sent"
                reactions: []  # Assuming the user had only one reaction which is now removed
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
components:
  securitySchemes:
    BearerAuth:
      description: Bearer authentication with a token identifier
      type: http
      scheme: bearer
      bearerFormat: userId
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema:
            type: object
            description: Error payload for invalid requests
            properties:
              description:
                type: string
                description: Human-readable error description
                pattern: '^[A-Za-z0-9_-]+$'
                minLength: 1
                maxLength: 64
            required:
              - description
    NotFoundResponse:    
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
  schemas:
...
    NotFound:
      description: Not found error payload
      type: object
      properties:
        message:
          description: message of the not found error
          type: string
          pattern: '^[A-Za-z0-9_-]+$'
          minLength: 1
          maxLength: 64
      required:
        - message
    UserId:
      type: string
      description: Unique identifier of the user
      pattern: '^usr_[A-Za-z0-9_-]{3,60}$'
      minLength: 7      # "usr_" (4) + minimo 3 = 7
      maxLength: 64
      example: 'usr_a1B2c3'
    ConversationId:
      type: string
      description: Unique identifier of the conversation
      pattern: '^conv_[A-Za-z0-9_-]{3,60}$'
      minLength: 8
      maxLength: 65
      example: 'conv_XyZ123'
    MessageId:
      type: string
      description: Unique identifier of the message
      pattern: '^msg_[A-Za-z0-9_-]{3,60}$'
      minLength: 7
      maxLength: 64
      example: 'msg_9aB'
    Message:
      type: object
      description: message sent by an user. Server requires at least one between text and media.
      properties:
        status:
          description: status of the message
          type: string
          enum: [seen, sent, deleted, delivered]
          pattern: '^(seen|sent|deleted|delivered)$'
          minLength: 4
          maxLength: 7
          readOnly: true
        media:
          $ref: '#/components/schemas/Media'
        reactions:
          type: array
          description: List of reactions added to this message
          items:
            $ref: '#/components/schemas/Reaction'
          minItems: 0
          maxItems: 100
          readOnly: true
        time:
          description: time that message was sent 
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
          minLength: 20
          maxLength: 20
          readOnly: true
        sender:
          description: ID of the sender
          allOf:
            - $ref: '#/components/schemas/UserId'
          readOnly: true

        messageId:
          description: Unique identifier for the message
          allOf:
            - $ref: '#/components/schemas/MessageId'
          readOnly: true
        kind:
          description: Discriminator for the message type
          type: string
          enum: [normal, forwarded]
        replyToId:
          allOf:
            - $ref: '#/components/schemas/MessageId'
          description: "id of the message being replied to (if this is a reply)"
        text:
          type: string
          description: message's text (optional but text or media are required)
          pattern: '^.+$'
          minLength: 1
          maxLength: 150
    Media:
      type: object
      description: media attached to a message or can be the only message
      properties:
        url:
          type: string
          format: url
          description: URL of image
          pattern: '^https?://.+$'
          minLength: 10
          maxLength: 2048
        filename:
          type: string
          description: file image name
          pattern: '^[A-Za-z0-9._-]{1,255}$'
          minLength: 1
          maxLength: 255
        mimeType:
          type: string
          description: image/png, image/jpeg
          pattern: '^image/(png|jpeg|jpg|gif)$'
          minLength: 10
          maxLength: 20
        size:
          type: integer
          description: image size in byte
      required:
        - url
        - mimeType
    Conversation: 
      type: object
      description: Conversation between users, either private or group
      discriminator:
        propertyName: type
        mapping:
          private: '#/components/schemas/PrivateConversation'
          group: '#/components/schemas/Group'
      properties:
        convId: 
          $ref: '#/components/schemas/ConversationId'
        type:
          type: string
          enum: [private, group]
          description: Discriminator to distinguish between 1-to-1 and groups
          default: private
        lastMessage:
          allOf:
            - $ref: '#/components/schemas/Message'
          description: Preview of the last message sent (for the list view)
          readOnly: true
        unreadCount:
           type: integer
           description: Number of unread messages in this conversation
           default: 0
           minimum: 0
           maximum: 99
           readOnly: true
        participantNames:
          type: array
          description: Display names of participants in the same order of participants array
          items:
            type: string
            minLength: 3
            maxLength: 16
            pattern: '^[A-Za-z0-9._-]{3,16}$'
          minItems: 1
          maxItems: 100
      required:
        - convId
        - type
    Reaction:
      type: object
      description: A reaction (emoji) added by a user to a message
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        emoji:
          type: string
          description: The emoji character
          pattern: '^.+$'
          minLength: 1
          maxLength: 4 
      required:
        - userId
        - emoji
    Group:
      allOf:
        - $ref: '#/components/schemas/Conversation'
        - type: object
          description: conversation with at least 2 users or more with option to set a name and a pic 
          properties:
            type: 
              type: string
              enum: [group]
              default: group
              description: Discriminator indicating that this conversation is a group
            groupName:
              type: string
              description: Group name
              default: "New_Group"
              pattern: '^[A-Za-z0-9\s_-]+$'
              minLength: 1
              maxLength: 64
            groupPhoto: 
              type: string
              format: url
              description: url of the group photo
              pattern: '^https?://.+$'
              minLength: 10
              maxLength: 2048
            participants:
              type: array
              description: IDs of the users that belong to the group
              items:
                $ref: '#/components/schemas/UserId'
              minItems: 1     
              maxItems: 100
            createdAt:
              type: string
              format: date-time
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
              minLength: 20
              maxLength: 20
              description: Group creation date
          required:
            - participants  


#    PrivateConversation:
#          allOf:
#            - $ref: '#/components/schemas/Conversation'
#            - type: object
#              properties:
#                type:
#                  type: string
#                  enum: [private]
#                  default: private
#                participants:
#                  type: array
#                  description: Exactly two users in a private chat
#                  items:
#                    $ref: '#/components/schemas/UserId'
#                  minItems: 2
#                  maxItems: 2
#              required:
#                - participants